add person to getattribute and split partial templates by scope
