+++ archive/templates/partial/person/name.html
@@ -1,0 +1,1 @@
+{{ name }}
