+++ archive/views/json/JsonGetAttributes.py
@@ -2,7 +2,7 @@ from django.views.generic import View
 from django.http import JsonResponse
 
 from .JsonHelper import JsonHelper
-from ...models import Image
+from ...models import Image, Person
 
 class GetAttribute:
   def __init__(self, object, attribute):
@@ -11,29 +11,34 @@ class GetAttribute:
 
 
 class JsonGetAttributeOfImage(JsonHelper):
-  def get(self, request, *args, **kwargs):
+  def get_object(self):
     ''' Get the Object 
         Try to get the object by pk, if not found try to get by slug
         If the object is not found, return an error message
     '''
     try:
-      self.object = Image.objects.get(pk=kwargs['pk']) if 'pk' in kwargs else Image.objects.get(slug=kwargs['slug'])
+      object = Image.objects.get(pk=self.kwargs['pk']) if 'pk' in self.kwargs else Image.objects.get(slug=self.kwargs['slug'])
     except Image.DoesNotExist:
       self.add_message('error', 'Image not found')
-      return self.return_response()
+      return False
     ''' If the object is deleted, return an error message '''
-    if self.object.is_deleted:
+    if object.is_deleted:
       self.add_message('error', 'Image is deleted')
-      return self.return_response
-    ''' Fetch the Attribute Name
-        The attribute name is the key to the value we want to retrieve
+      return False
+    return object
+  
+  
+
+class JsonGetAttributeOfPerson(JsonHelper):
+  def get_object(self):
+    ''' Get the Object 
+        Try to get the object by pk, if not found try to get by slug
+        If the object is not found, return an error message
     '''
-    self.attribute = self.get_attribute()
-    if not self.attribute:
-      return self.return_response()
-    ''' Get the Value of the Attribute '''
-    value = getattr(self.object, self.attribute)
-    value = value.all() if hasattr(value, 'all') else value
-    ''' Add the Value to the Response '''
-    self.add_payload(value)
-    return self.return_response()
+    try:
+      object = Person.objects.get(pk=self.kwargs['pk']) if 'pk' in self.kwargs else Person.objects.get(slug=self.kwargs['slug'])
+    except Person.DoesNotExist:
+      self.add_message('error', 'Person not found')
+      return False
+    return object
+
